% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.import_fragments}
\alias{.import_fragments}
\title{Importing fragment files}
\usage{
.import_fragments(file, withIndex = TRUE)
}
\description{
Importing fragment files
}
\examples{

con <- TENxFile("~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5")
sce <- import(con)
fr <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_atac_fragments.tsv.gz"
fra <- .import_fragments(fr)

}
\keyword{internal}
