% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.import_peak_anno}
\alias{.import_peak_anno}
\title{Importing peak-annotation files}
\usage{
.import_peak_anno(file)
}
\description{
Importing peak-annotation files
}
\examples{

fi <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_atac_peak_annotation.tsv"
pa <- .import_peak_anno(fi)

con <- TENxFile("~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5")
sce <- import(con)
annotation(sce, name = "peak_annotation") <- pa

}
\keyword{internal}
