% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShellExperiment-class.R
\docType{class}
\name{ShellExperiment-class}
\alias{ShellExperiment-class}
\alias{.ShellExperiment}
\title{The ShellExperiment class for temporary representations of H5 files}
\description{
The class represents a deconstructed \code{SingleCellExperiment}
class for inspecting several aspects of the \code{TENxH5} file class. It
inherits the methods from the \code{TENxH5} class. See the \code{TENxH5-class}
documentation.
}
\section{Slots}{

\describe{
\item{\code{assays}}{Assays_OR_NULL; Assay class input similar to that in the
\code{SummarizedExperiment} constructor (default \code{NULL})}

\item{\code{rowRanges}}{GenomicRanges_OR_GRangesList; a ranged based representation
of the genomic data within the file}
}}

