% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShellExperiment-class.R
\name{ShellExperiment}
\alias{ShellExperiment}
\title{ShellExperiment constructor function}
\usage{
ShellExperiment(con)
}
\arguments{
\item{con}{character(1) A path to an H5 file}
}
\description{
The \code{ShellExperiment} function takes a version 2 or 3 HDF5
dataset typically obtained via the 10x website and provides an intermediate
representation of the data within the file. Supported methods include
\code{rowRanges}, \code{dimnames}, and \code{dim}.
}
\examples{

h5f <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5"
con <- TENxFile(h5f)
ShellExperiment(con)

## Alternatively
mat <- HDF5Array::TENxMatrix(h5f, "matrix")
hub <- ExperimentHub::ExperimentHub()
fname <- hub[["EH1039"]]
x5 <- TENxH5(fname, group = "mm10", version = "2")

}
