% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxFile-class.R
\name{TENxFile}
\alias{TENxFile}
\title{TENxFile constructor function}
\usage{
TENxFile(resource, group = c("matrix", "outs"), version = c("3", "2"))
}
\description{
TENxFile constructor function
}
\examples{

h5f <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5"
con <- TENxFile(h5f)
import(con)

}
