% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxFileList-class.R
\docType{class}
\name{TENxFileList-class}
\alias{TENxFileList-class}
\alias{.TENxFileList}
\alias{path,TENxFileList-method}
\title{TENxFileList: A list-like representation for TENxFiles}
\usage{
\S4method{path}{TENxFileList}(object, ...)
}
\description{
This class was designed to mainly handle tarballs from
10X Genomics. The typical file extension for these tarballs is \code{.tar.gz}.
}
\details{
These tarballs usually contain three files:
\enumerate{
\item \code{matrix.mtx.gz} - the counts matrix
\item \code{features.tsv.gz} - row metadata usually represented as \code{rowData}
\item \code{barcodes.tsv.gz} - column names corresponding to cell barcode identifiers
}
}
\section{Functions}{
\itemize{
\item \code{path,TENxFileList-method}: Obtain file paths for all files in the object
as a vector
}}

\section{Slots}{

\describe{
\item{\code{listData}}{list() The data in list format}

\item{\code{extension}}{character() A vector of file extensions for each file}

\item{\code{compressed}}{logical(1) Whether the file is compressed as \code{.tar.gz}}
}}

