% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxH5-class.R
\docType{class}
\name{TENxH5-class}
\alias{TENxH5-class}
\alias{.TENxH5}
\alias{rowData,TENxH5-method}
\alias{dim,TENxH5-method}
\alias{dimnames,TENxH5-method}
\alias{genome,TENxH5-method}
\alias{rowRanges,TENxH5-method}
\alias{import,TENxH5,ANY,ANY-method}
\title{TENxH5: The HDF5 file representation class for 10X Data}
\usage{
\S4method{rowData}{TENxH5}(x, use.names = TRUE, ...)

\S4method{dim}{TENxH5}(x)

\S4method{dimnames}{TENxH5}(x)

\S4method{genome}{TENxH5}(x)

\S4method{rowRanges}{TENxH5}(x, ...)

\S4method{import}{TENxH5,ANY,ANY}(con, format, text, ...)
}
\description{
This class is designed to work with 10x Single Cell datasets.
It was developed using the PBMC 3k 10X dataset from the CellRanger v2
pipeline.
}
\details{
The data version "3" mainly includes a "matrix" group and "interval"
information within the file. Version "2" data does not include
ranged-based information and has a different directory structure compared
to version "3". See the internal \code{data.frame}: \code{TENxIO:::gene.meta} for
a map of fields and their corresponding file locations within the H5 file.
This map is used to create the \code{rowData} structure from the file.
}
\section{Methods (by generic)}{
\itemize{
\item \code{rowData}: Generate the rowData ad hoc from a TENxH5 file

\item \code{dim}: Get the dimensions of the data as stored in the file

\item \code{dimnames}: Get the dimension names from the file

\item \code{genome}: Read genome string from file

\item \code{rowRanges}: Read interval data and represent as GRanges

\item \code{import}: Import TENxH5 data as a SingleCellExperiment
}}

\section{Slots}{

\describe{
\item{\code{version}}{character(1) There are currently two recognized versions
associated with 10X data, either version "2" or "3". See details for more
information.}

\item{\code{group}}{character(1) The HDF5 group embedded within the file structure,
this is usually either the "matrix" or "outs" group but other groups are
supported as well.}
}}

