% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxMTX-class.R
\docType{class}
\name{TENxMTX-class}
\alias{TENxMTX-class}
\alias{.TENxMTX}
\title{TENxMTX: The Matrix Market representation class for 10X Data}
\description{
This class is designed to work with 10x MTX datasets,
particularly from the multiome pipelines.
}
\details{
The \code{TENxMTX} class is a straightforward implementation that allows
the user to import a Matrix Market file format using \code{Matrix::readMM}.
Currently, it only supports return types of \code{dgCMatrix}. To request other
formats, please open an issue on GitHub.
}
\section{Slots}{

\describe{
\item{\code{compressed}}{logical(1) Whether or not the file is in compressed format,
usually gzipped (\code{.gz}).}
}}

