% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxPeaks-class.R
\docType{class}
\name{TENxPeaks-class}
\alias{TENxPeaks-class}
\alias{.TENxPeaks}
\title{TENxPeaks: The class to represent 10x Peaks files}
\description{
This class is designed to work with the files denoted with "peak_annotation"
in the file name. These are usually produced as tab separated value files,
i.e., \code{.tsv}.
}
\details{
This class is a straightforward class for handling peak data. It can
be used in conjunction with the \code{annotation} method on a
\code{SingleCellExperiment} to add peak information to the experiment. The
ranged data is represented as a \code{GRanges} class object.
}
