% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxPeaks-class.R
\name{TENxPeaks}
\alias{TENxPeaks}
\title{Import 10x peak annotation files from 10x}
\usage{
TENxPeaks(resource, ...)
}
\arguments{
\item{resource}{character(1) The path to the file}

\item{...}{Additional inputs to the low level class generator functions}
}
\value{
A \code{TENxPeak} class object
}
\description{
This constructor function is designed to work with the files denoted with
"peak_annotation" in the file name. These are usually produced as tab
separated value files, i.e., \code{.tsv}.
}
\details{
The output class allows handling of peak data. It can be used in
conjunction with the \code{annotation} method on a \code{SingleCellExperiment} to add
peak information to the experiment. The ranged data is represented as a
\code{GRanges} class object.
}
\examples{

fi <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_atac_peak_annotation.tsv"
peak_file <- TENxPeaks(fi)
peak_anno <- import(pa)

con <- TENxFile("~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5")
sce <- import(con)
annotation(sce, name = "peak_annotation") <- peak_file

}
