% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxFileList-class.R
\docType{class}
\name{TSVFile-class}
\alias{TSVFile-class}
\alias{.TSVFile}
\alias{import,TSVFile,ANY,ANY-method}
\title{TSVFile: A class to represent 10x tab separated values files}
\usage{
\S4method{import}{TSVFile,ANY,ANY}(con, format, text, ...)
}
\arguments{
\item{con}{ The connection from which data is loaded or to which data
    is saved. If this is a character vector, it is assumed to be a
    filename and a corresponding file connection is created and then
    closed after exporting the object. If a \code{\linkS4class{BiocFile}}
    derivative, the data is loaded from or saved to the underlying
    resource. If missing, the function will return the output as a
    character vector, rather than writing to a connection.
  }

\item{format}{ The format of the output. If missing and \code{con} is
    a filename, the format is derived from the file extension. This
    argument is unnecessary when \code{con} is a derivative of
    \code{\linkS4class{BiocFile}}.
  }

\item{text}{ If \code{con} is missing, this can be a character vector
    directly providing the string data to import. }

\item{...}{ Parameters to pass to the format-specific method.
  }
}
\value{
A \code{TSVFile} class object; a \code{tibble} for the import method
}
\description{
This class is general purpose for reading in tabular data from the
10x Genomics website with the \code{.tsv} file extension. The class also supports
compressed files, i.e., those with the \code{.tsv.gz} extension.
}
\section{Methods (by generic)}{
\itemize{
\item \code{import(con = TSVFile, format = ANY, text = ANY)}: General import function for \code{tsv} files from 10x;
using \code{readr::read_tsv} and returning a \code{tibble} representation

}}
\keyword{internal}
