% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.import_fragments}
\alias{.import_fragments}
\title{Importing fragment files}
\usage{
.import_fragments(
  file,
  roi = GRanges(),
  yieldSize = 200,
  index = paste0(file, ".tbi")
)
}
\arguments{
\item{file}{character(1) The file path to the fragments resource, usually a
compressed tabix file with extension \code{.tsv.gz}.}

\item{roi}{GRanges() A GRanges indicating the regions of interest. This
get sent to \code{RSamtools} as the \code{param} input.}

\item{yieldSize}{numeric() The number of records to read by default, 200
records will be imported. A warning will be emmitted if not modified.}

\item{index}{character(1) The index file usually ending in \code{.tbi}. It
should be found in the same location as the \code{file} input}
}
\description{
Importing fragment files
}
\examples{

con <- TENxFile("~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5")
sce <- import(con)

fr <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_atac_fragments.tsv.gz"
fra <- .import_fragments(fr)

}
\keyword{internal}
