% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxFileList-class.R
\name{TENxFileList}
\alias{TENxFileList}
\title{TENxFileList: Represent groups of files from 10X Genomic}
\usage{
TENxFileList(..., version, compressed = FALSE)
}
\arguments{
\item{...}{A single file path, named arguments corresponding to file paths,
or a list of named file paths}

\item{version}{character(1) The version in the tarball. See details.}

\item{compressed}{logical(1) Whether or not the file provided is compressed,
usually as \code{tar.gz} (default FALSE)}
}
\value{
Either a \code{SingleCellExperiment} or a list of imported data
}
\description{
This constructor function is meant to handle \code{.tar.gz} tarball
files from 10X Genomics.
}
\details{
These tarballs usually contain three files:
\enumerate{
\item \code{matrix.mtx.gz} - the counts matrix
\item \code{features.tsv.gz} - row metadata usually represented as \code{rowData}
\item \code{barcodes.tsv.gz} - column names corresponding to cell barcode
identifiers
If all above files are in the tarball, the import method will provide a
\code{SingleCellExperiment}. Otherwise, a simple list of imported data is given.
Note that version "3" uses 'features.tsv.gz' and version "2" uses
'genes.tsv.gz'. If known, indicate the version argument in the \code{TENxFileList}
constructor function.
}
}
\examples{

fl <- system.file(
    "extdata", "pbmc_granulocyte_sorted_3k_ff_bc_ex_matrix.tar.gz",
    package = "TENxIO", mustWork = TRUE
)

import(TENxFileList(fl))

}
