% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxH5-class.R
\docType{class}
\name{TENxH5-class}
\alias{TENxH5-class}
\alias{.TENxH5}
\alias{rowData,TENxH5-method}
\alias{dim,TENxH5-method}
\alias{dimnames,TENxH5-method}
\alias{genome,TENxH5-method}
\alias{rowRanges,TENxH5-method}
\alias{import,TENxH5,ANY,ANY-method}
\alias{show,TENxH5-method}
\title{TENxH5: The HDF5 file representation class for 10X Data}
\usage{
\S4method{rowData}{TENxH5}(x, use.names = TRUE, ...)

\S4method{dim}{TENxH5}(x)

\S4method{dimnames}{TENxH5}(x)

\S4method{genome}{TENxH5}(x)

\S4method{rowRanges}{TENxH5}(x, ...)

\S4method{import}{TENxH5,ANY,ANY}(con, format, text, ...)

\S4method{show}{TENxH5}(object)
}
\arguments{
\item{...}{ Parameters to pass to the format-specific method.
  }

\item{con}{ The connection from which data is loaded or to which data
    is saved. If this is a character vector, it is assumed to be a
    filename and a corresponding file connection is created and then
    closed after exporting the object. If a \code{\linkS4class{BiocFile}}
    derivative, the data is loaded from or saved to the underlying
    resource. If missing, the function will return the output as a
    character vector, rather than writing to a connection.
  }

\item{format}{ The format of the output. If missing and \code{con} is
    a filename, the format is derived from the file extension. This
    argument is unnecessary when \code{con} is a derivative of
    \code{\linkS4class{BiocFile}}.
  }

\item{text}{ If \code{con} is missing, this can be a character vector
    directly providing the string data to import. }

\item{object}{A \code{TENxH5} class object}
}
\description{
This class is designed to work with 10x Single Cell datasets.
It was developed using the PBMC 3k 10X dataset from the CellRanger v2
pipeline.
}
\details{
The data version "3" mainly includes a "matrix" group and "interval"
information within the file. Version "2" data does not include
ranged-based information and has a different directory structure compared
to version "3". See the internal \code{data.frame}: \code{TENxIO:::gene.meta} for
a map of fields and their corresponding file locations within the H5 file.
This map is used to create the \code{rowData} structure from the file.
}
\section{Methods (by generic)}{
\itemize{
\item \code{rowData}: Generate the rowData ad hoc from a TENxH5 file

\item \code{dim}: Get the dimensions of the data as stored in the file

\item \code{dimnames}: Get the dimension names from the file

\item \code{genome}: Read genome string from file

\item \code{rowRanges}: Read interval data and represent as GRanges

\item \code{import}: Import TENxH5 data as a SingleCellExperiment

\item \code{show}: Display a snapshot of the contents within a TENxH5 file
before import
}}

\section{Slots}{

\describe{
\item{\code{version}}{character(1) There are currently two recognized versions
associated with 10X data, either version "2" or "3". See details for more
information.}

\item{\code{group}}{character(1) The HDF5 group embedded within the file structure,
this is usually either the "matrix" or "outs" group but other groups are
supported as well.}

\item{\code{ranges}}{character(1) The HDF5 internal folder location embedded within
the file that points to the ranged data information, e.g.,
"/features/interval".}
}}

