% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxH5-class.R
\name{TENxH5}
\alias{TENxH5}
\alias{TENxH5-class}
\title{Import H5 files from 10X}
\usage{
TENxH5(resource, version, group, ...)
}
\description{
Import H5 files from 10X
}
\examples{

h5f <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5"
con <- TENxH5(h5f)
import(con)

}
