% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxH5-class.R
\name{TENxH5}
\alias{TENxH5}
\title{TENxH5: Represent H5 files from 10X}
\usage{
TENxH5(resource, version, group, ...)
}
\arguments{
\item{resource}{character(1) The path to the file}

\item{version}{character(1) There are currently two recognized versions
associated with 10X data, either version "2" or "3". See details for more
information.}

\item{group}{character(1) The HDF5 group embedded within the file structure,
this is usually either the "matrix" or "outs" group but other groups are
supported as well (e.g., "mm10").}

\item{...}{Additional inputs to the low level class generator functions}
}
\description{
This constructor function was developed using the PBMC 3K dataset from 10X
Genomics (version 3). Other versions are supported and input arguments
\code{version} and \code{group} can be overridden.
}
\details{
The various \code{TENxH5} methods including \code{rowData} and \code{rowRanges},
provide a snapshot of the data using a length 12 head and tail subset for
efficiency. In contrast, methods such as \code{dimnames} and \code{dim} give
a full view of the dimensions of the data. The \code{show} method provides
relevant information regarding the dimensions of the data including
metadata such as \code{rowData} and "Type" column, if available. The term
"projection" refers to the data class that will be provided once the
data file is \code{import}ed.

The data version "3" mainly includes a "matrix" group and "interval"
information within the file. Version "2" data does not include
ranged-based information and has a different directory structure compared
to version "3". See the internal \code{data.frame}: \code{TENxIO:::gene.meta} for
a map of fields and their corresponding file locations within the H5 file.
This map is used to create the \code{rowData} structure from the file.
}
\examples{

h5f <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5"
con <- TENxH5(h5f)
import(con)

}
