% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxMTX-class.R
\docType{class}
\name{TENxMTX-class}
\alias{TENxMTX-class}
\alias{.TENxMTX}
\title{TENxMTX: The Matrix Market representation class for 10X Data}
\description{
TENxMTX: The Matrix Market representation class for 10X Data
}
