% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENxPeaks-class.R
\name{TENxPeaks}
\alias{TENxPeaks}
\title{Importing peak-annotation files}
\usage{
TENxPeaks(resource, ...)
}
\description{
Importing peak-annotation files
}
\examples{

fi <- "~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_atac_peak_annotation.tsv"
pa <- TENxPeaks(fi)
import(pa)

con <- TENxFile("~/data/10x/pbmc_3k/pbmc_granulocyte_sorted_3k_filtered_feature_bc_matrix.h5")
sce <- import(con)
annotation(sce, name = "peak_annotation") <- pa

}
